+++ lib/src/rust/frb_generated.web.dart
@@ -173,6 +173,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   List<SemanticChunk> dco_decode_list_semantic_chunk(dynamic raw);
 
   @protected
+  List<SourceEntry> dco_decode_list_source_entry(dynamic raw);
+
+  @protected
   List<StructuredChunk> dco_decode_list_structured_chunk(dynamic raw);
 
   @protected
@@ -217,6 +220,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   SemanticChunk dco_decode_semantic_chunk(dynamic raw);
+
+  @protected
+  SourceEntry dco_decode_source_entry(dynamic raw);
 
   @protected
   SourceStats dco_decode_source_stats(dynamic raw);
@@ -417,6 +423,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   );
 
   @protected
+  List<SourceEntry> sse_decode_list_source_entry(SseDeserializer deserializer);
+
+  @protected
   List<StructuredChunk> sse_decode_list_structured_chunk(
     SseDeserializer deserializer,
   );
@@ -471,6 +480,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   SemanticChunk sse_decode_semantic_chunk(SseDeserializer deserializer);
+
+  @protected
+  SourceEntry sse_decode_source_entry(SseDeserializer deserializer);
 
   @protected
   SourceStats sse_decode_source_stats(SseDeserializer deserializer);
@@ -734,6 +746,12 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
   );
 
   @protected
+  void sse_encode_list_source_entry(
+    List<SourceEntry> self,
+    SseSerializer serializer,
+  );
+
+  @protected
   void sse_encode_list_structured_chunk(
     List<StructuredChunk> self,
     SseSerializer serializer,
@@ -798,6 +816,9 @@ abstract class RustLibApiImplPlatform extends BaseApiImpl<RustLibWire> {
 
   @protected
   void sse_encode_semantic_chunk(SemanticChunk self, SseSerializer serializer);
+
+  @protected
+  void sse_encode_source_entry(SourceEntry self, SseSerializer serializer);
 
   @protected
   void sse_encode_source_stats(SourceStats self, SseSerializer serializer);